# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GraphDB API",
    "description": "GraphDB network service layer (HTTP/gRPC)",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/schema/spaces/{name}/edge-types": {
      "get": {
        "tags": [
          "WebSchema"
        ],
        "summary": "List all edge types in a space",
        "operationId": "get_api_v1_schema_spaces_name_edge_types",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Space name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Edge type list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "WebSchema"
        ],
        "summary": "Create a new edge type",
        "operationId": "post_api_v1_schema_spaces_name_edge_types",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Space name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Edge type created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/v1/schema/spaces/{name}/edge-types": {
      "get": {
        "tags": [
          "Schema"
        ],
        "summary": "List all edge types",
        "operationId": "get_v1_schema_spaces_name_edge_types",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Space name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Edge type list",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Schema"
        ],
        "summary": "Creating Edge Types",
        "operationId": "post_v1_schema_spaces_name_edge_types",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Space name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEdgeTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Edge type created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponse_Value": {
        "type": "object",
        "description": "Standard API response wrapper",
        "required": [
          "success"
        ],
        "properties": {
          "data": {},
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "CreateEdgeTypeRequest": {
        "type": "object",
        "description": "Create edge type request",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyDef"
            }
          }
        }
      },
...
      "PropertyDef": {
        "type": "object",
        "description": "Property definition",
        "required": [
          "name",
          "data_type",
          "nullable"
        ],
        "properties": {
          "data_type": {
            "type": "string"
          },
          "default_value": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "nullable": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}